customize some code and component
