+++ src/App.css
@@ -43,7 +43,6 @@ body {
   border-radius: 1px;
   border: none;
   outline: none;
-  cursor: pointer;
 }
 
 .green_gradient_text {
