+++ src/page/dashboard/Student/certificates/pdf/PDF.jsx
@@ -8,10 +8,8 @@ function PDF({anotherName}) {
         <Document>
             <Page size="A4">
                 <View style={{ position: 'relative', overflow: 'hidden' }}>
-                    {/* <div id='certificate'> */}
-                    <Text style={{ position: "absolute", top: '380px', right: 0, left: 100, zIndex: 1, fontSize: '60px', textTransform: 'capitalize', fontWeight: 'bold' }} >{anotherName}</Text>
-                    <Image src='https://i.ibb.co/3yMGWhT/certificate-new.png' style={{ width: '600', height: '800', zIndex: 3 }} />
-                    {/* </div> */}
+                    <Text style={{ position: "absolute", top: '395px', right: 0, left: 110, zIndex: 1, fontSize: '60px', textTransform: 'capitalize', fontWeight: 'bold' }} >{anotherName}</Text>
+                    <Image src='https://i.ibb.co/N6gZBXk/certificate.png' style={{ width: '600', height: '800', zIndex: 3 }} />
                 </View>
             </Page>
         </Document>
